refactor: move email code fetch logic
